# backend/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Viv Scanner - Firmware Security Analysis Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /scanner/start:
    post:
      operationId: startScan
      tags: [scanner]
      summary: Start a firmware scan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanStartInput"
      responses:
        "201":
          description: Scan started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanResult"
  /scanner/results/{scanId}:
    get:
      operationId: getScanResults
      tags: [scanner]
      summary: Get scan results for a scan execution (by scanId)
      parameters:
        - name: scanId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scan results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScanResult"
  /scanner/files/{scanId}:
    get:
      operationId: getExtractedFiles
      tags: [scanner]
      summary: Get extracted files for a scan execution (by scanId)
      parameters:
        - name: scanId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Extracted file list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExtractedFile"
  /scanner/binary/{id}:
    post:
      operationId: runBinaryAnalysis
      tags: [scanner]
      summary: Run binary analysis on extracted binaries
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BinaryAnalysisInput"
      responses:
        "200":
          description: Binary analysis results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BinaryAnalysisResult"
...
  /reports/pdf/{id}:
    get:
      operationId: getPdfReport
      tags: [reports]
      summary: Get PDF report for a firmware scan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: PDF report info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportMeta"
  /reports/ai-summary/{id}:
    get:
      operationId: getAiSummary
      tags: [reports]
      summary: Get AI-generated summary and risk analysis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: AI summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiReport"
  /reports/sbom/{id}:
    get:
      operationId: getSbomReport
      tags: [reports]
      summary: Get SBOM metadata and component list for a firmware scan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: SBOM metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SbomReport"
  /reports/sbom/{id}/download/{format}:
    get:
      operationId: getSbomDownload
      tags: [reports]
      summary: Download SBOM file in the requested format
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: format
          in: path
          required: true
          schema:
            type: string
            enum: [cyclonedx, spdx, csv]
      responses:
        "200":
          description: SBOM file download
          content:
            "*/*":
              schema:
                type: string
                format: binary
...
components:
  schemas:
...
    ScanResult:
      type: object
      required: [id, firmwareId, startedAt, status, progress]
      properties:
        id:
          type: integer
        firmwareId:
          type: integer
        startedAt:
          type: string
        completedAt:
          type: ["string", "null"]
        status:
          type: string
          enum: [running, completed, failed]
        progress:
          type: integer
        totalFiles:
          type: ["integer", "null"]
        vulnerabilitiesFound:
          type: ["integer", "null"]
        riskLevel:
          type: ["string", "null"]
          enum: [critical, high, medium, low, null]
    ScanStartInput:
      type: object
      required: [firmwareId]
      properties:
        firmwareId:
          type: integer
        deepScan:
          type: boolean
    ExtractedFile:
      type: object
      required: [id, firmwareId, path, type, size]
      properties:
        id:
          type: integer
        firmwareId:
          type: integer
        path:
          type: string
        type:
          type: string
        size:
          type: integer
        permissions:
          type: ["string", "null"]
        isSuspicious:
          type: boolean
    BinaryAnalysisInput:
      type: object
      required: [filePath]
      properties:
        filePath:
          type: string
    BinaryAnalysisResult:
      type: object
      required: [firmwareId, filePath, strings, urls, credentials, debugPaths]
      properties:
        firmwareId:
          type: integer
        filePath:
          type: string
        strings:
          type: array
          items:
            type: string
        urls:
          type: array
          items:
            type: string
        credentials:
          type: array
          items:
            type: string
        debugPaths:
          type: array
          items:
            type: string
...
    ReportMeta:
      type: object
      required: [firmwareId, generatedAt, downloadUrl, fileSize]
      properties:
        firmwareId:
          type: integer
        generatedAt:
          type: string
        downloadUrl:
          type: string
        fileSize:
          type: integer
        scanId:
          type: ["integer", "null"]
    AiReport:
      type: object
      required: [firmwareId, summary, riskLevel, keyFindings, recommendations, generatedAt]
      properties:
        firmwareId:
          type: integer
        summary:
          type: string
        riskLevel:
          type: string
          enum: [critical, high, medium, low]
        keyFindings:
          type: array
          items:
            type: string
        recommendations:
          type: array
          items:
            type: string
        generatedAt:
          type: string
        exploitProbability:
          type: ["number", "null"]
    SbomComponent:
      type: object
      required: [name, version, type, path, source]
      properties:
        name:
          type: string
        version:
          type: string
        type:
          type: string
        path:
          type: string
        source:
          type: string
    SbomReport:
      type: object
      required: [firmwareId, generatedAt, componentCount, downloadUrls, components]
      properties:
        firmwareId:
          type: integer
        generatedAt:
          type: string
        componentCount:
          type: integer
        downloadUrls:
          type: object
          required: [cyclonedx, spdx, csv]
          properties:
            cyclonedx:
              type: string
            spdx:
              type: string
            csv:
              type: string
        components:
          type: array
          items:
            $ref: "#/components/schemas/SbomComponent"
...